CS 350 Module seven final Project Thermostat Lab
Objective
The primary aim of this module was to develop a smart thermostat system based on Raspberry Pi with other real-time components such as the LCD display, LED indicators, input buttons, and temperature and humidity sensor (AHT20). The solution is a basic heating-cooling controller, enhanced with visual feedback and user interaction by a series of buttons.