CS_465_Project_wireframe_class_diagram
This a  project architecture class diagram is a UML (Unified Modeling Language) diagram that provides a static, structural view of a software system, illustrating its classes, their attributes, and operations, as well as the relationships between them, such as inheritance and association. It serves as a blueprint for the system's design, capturing the fundamental organization and dependencies of its components, which aids in communication, planning, and development for architects, developers, and other stakeholders.